Spec current behavior wrt. subsequences
The question is, given an input of [7,8,9,4,5,6,1,2,3,4] and N=3, what
should the program's output be?

One approach is to take all monotonic sequences and take 3 longest,
which would be (skipping all sequences of size 2 or smaller):

  [7,8,9], [4,5,6], [1,2,3], [1,2,3,4]

and take 3 longest ones, resulting in [[4,5,6],[1,2,3],[1,2,3,4]].

The other approach is take the fact that [1,2,3] is a sub-sequence of
[1,2,3,4] and to only count the latter, resulting in [[7,8,9],[4,5,6],[1,2,3,4]]

Server currently takes the first approach, which is documented in this
test.